# docs/swagger.yaml
basePath: /gestrym-storage
info:
  contact: {}
  description: API para el manejo de entrenamientos.
  title: Gestrym API Storage
  version: "1.0"
paths:
  /internal/files/{id}:
    delete:
      description: Elimina un archivo del storage y de la base de datos
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Eliminar archivo
      tags:
      - Storage
  /internal/files/collection:
    get:
      description: Retorna los archivos asociados a una colección específica
      parameters:
      - description: Collection ID
        in: query
        name: collectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Obtener archivos por ID de colección
      tags:
      - Storage
  /internal/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Sube archivos de forma concurrente, valida tipo y tamaño
      parameters:
      - description: Collection ID to relate (optional, will generate one if empty)
        in: formData
        name: collectionId
        type: string
      - description: Service of origin (e.g., 'users', 'exercises')
        in: formData
        name: service
        type: string
      - description: Archivos a subir
        in: formData
        name: files
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Subir uno o varios archivos
      tags:
      - Storage
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"